- Bills module. Basic; - Operations module. Basic; - Categories module. Basic;

+++ apps/finapp/backend/src/app/modules/base-service.ts
@@ -6,11 +6,11 @@ export class BaseService<T, U extends GetParamsData> {
 	protected repository: Repository<T>;
 	protected relations: string[];
 	protected entityNotFoundMessage: string;
-	
+
 	public async getAll(): Promise<T[]> {
 		return await this.repository.find();
 	}
-	
+
 	public async getByID(id: number): Promise<T> {
 		const entity = await this.repository.findOne({where: {id}, relations: this.relations});
 		if (entity) {
@@ -18,42 +18,45 @@ export class BaseService<T, U extends GetParamsData> {
 		}
 		throw new HttpException(this.entityNotFoundMessage, HttpStatus.NOT_FOUND);
 	}
-	
+
 	public async getBy(@Param() paramsData: U): Promise<T> {
 		try {
 			const requestObject: FindOneOptions<T> = {
 				where: {...paramsData.params}
 			};
-			
+
 			if (paramsData.withRelations) {
 				requestObject.relations = this.relations;
 			}
-			
+
 			const entity = await this.repository.findOne(requestObject);
 			if (entity) {
 				return entity;
 			}
-			
+
 			if (paramsData.checkOnly) {
 				return;
 			}
-			
+
 			throw new HttpException(this.entityNotFoundMessage, HttpStatus.NOT_FOUND);
 		}
 		catch (e) {
 			throw new Error(e);
 		}
 	}
-	
+
 	public async delete(id: number): Promise<any> {
-		try {
-			const entity = await this.repository.findOne({where: {id}});
-			await this.repository.remove(entity);
-			return {status: HttpStatus.OK, statusText: 'Deleted successfully'};
-		}
-		catch (e) {
-			throw new Error(e);
-		}
+    const entity = await this.repository.findOne({where: {id}});
+    if (entity) {
+      try {
+        await this.repository.remove(entity);
+        return {status: HttpStatus.OK, statusText: 'Deleted successfully'};
+      }
+      catch (e) {
+        throw new Error(e);
+      }
+    }
+    throw new HttpException(this.entityNotFoundMessage, HttpStatus.NOT_FOUND);
 	}
 }
